+++ src/main/java/senac/domain/model/Moedas.java
@@ -18,7 +18,7 @@ public class Moedas {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "codmoeda")
-    private Long codMoeda;
+    private Integer codMoeda;
 
     @Column(name = "po", length = 45, columnDefinition = "VARCHAR(45) COMMENT 'peças de ouro'")
     private String po;
